+++ JavaBackup/AllProjects/src/games/generic/controlModel/gObj/MovingObject.java
@@ -5,10 +5,10 @@ import games.generic.controlModel.GModality;
 public interface MovingObject extends TimedObject, ObjectInSpace {
 
 	/** Progress the object's movement by the milliseconds elapsed */
-	public void move(long milliseconds);
+	public void move(int milliseconds);
 
 	@Override
-	public default void act(GModality modality, long milliseconds) {
+	public default void act(GModality modality, int milliseconds) {
 		move(milliseconds);
 	}
 }
